backend/requirements.txt
# requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Multipart Form Data
python-multipart>=0.0.6

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.0

# Email Validation (required by pydantic EmailStr)
email-validator>=2.0.0
